Remove use of Decimal and instead use float consistently

+++ app/domains/betting/entities.py
@@ -1,6 +1,5 @@
 from dataclasses import dataclass
 from datetime import datetime
-from decimal import Decimal
 from typing import Optional
 
 from ..predictions.entities import Prediction
@@ -11,17 +10,17 @@ from ..shared.value_objects import Odds
 class BettingOpportunity:
     prediction: Prediction
     recommended_outcome: str  # 'H', 'D', 'A'
-    edge: Decimal
-    fair_odds: Decimal
+    edge: float
+    fair_odds: float
     market_odds: Odds
-    expected_value: Decimal
-    kelly_fraction: Optional[Decimal] = None
+    expected_value: float
+    kelly_fraction: Optional[float] = None
 
 
 @dataclass
 class BettingResult:
     opportunity: BettingOpportunity
-    stake: Decimal
+    stake: float
     outcome: str
-    profit: Decimal
+    profit: float
     placed_at: datetime
